Update luis-rojas-article.md
@codephobe 

I am submitting my current edits as a pull request because I believe it may be easier for you and because I did not need to leave any comments. Your changes were great and seem to be just what this needed. 

With these edits, I believe it is ready to move to the Profiles editor to prepare for publishing. Here are the instructions for doing that: https://study.moderndeveloper.com/writing-editing-and-publishing-your-developer-of-the-week-article/

Great job! I have enjoyed working with you on this.